+++ src/components/atoms/Animated/Blinking.stories.tsx
@@ -1,0 +1,8 @@
+import Blinking from './Blinking';
+
+export default {
+  title: 'Blinking',
+  component: Blinking,
+};
+
+export const Default = {};
