--- src/components/atoms/Animated/stories.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react-native';
-import Blinking, { Props } from './Blinking';
-
-const props = (): Props => ({});
-
-storiesOf('atoms/Animated', module).add(
-  'Blinking',
-  () => <Blinking {...props()} />,
-  { loki: { skip: true } }
-);
